detector/requirements.txt
pyyaml==6.0.1
requests==2.31.0
flask==3.0.0
psutil==5.9.6
collections-extended==2.0.2
